refactor(auth): unify namespace for AuthUser related classes
