+++ services/auth/src/Auth.Persistence/Db/AuthDbContext.cs
@@ -1,4 +1,4 @@
-using Auth.Domain.Entities;
+using Auth.Domain.AuthUser;
 using Microsoft.EntityFrameworkCore;
